chore: Re-design of acquisition window
close #214
